backend/requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
aiohttp>=3.8.4
httpx>=0.24.0

# Language model and vectorstore
langchain>=0.0.267
langchain-google-genai>=0.0.1
google-generativeai>=0.3.0
faiss-cpu>=1.7.4  # Use faiss-gpu for GPU support

# PDF processing
PyPDF2>=3.0.0
pdfminer.six>=20221105
pdf2image>=1.16.3

# Image processing
Pillow>=9.5.0

# Memory and caching
redis>=4.5.5
diskcache>=5.4.0

# Gradio client
gradio-client>=0.2.7

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.2
tiktoken>=0.4.0  # For token counting
numpy>=1.24.3

# Optional but recommended
aiofiles>=23.1.0
asyncio>=3.4.3
tqdm>=4.65.0
colorama>=0.4.6  # For colored terminal output
loguru>=0.7.0  # Better logging
typing-extensions>=4.5.0

fastapi

uvicorn

google-generativeai

langchain-google-genai

langchain

PyPDF2

python-dotenv

faiss-cpu

pydantic

gradio_client

python-multipart

langchain_community
